pyproject.toml
[tool.poetry]
name = "LinkCut"
version = "0.1.1"
description = ""
authors = ["ahzees <matviy.stets@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.10"
psycopg2 = "^2.9.5"
django = "^4.1.7"
django-dotenv = "^1.4.2"
commitizen = "^2.42.0"
djangorestframework = "^3.14.0"
django-rest-auth = "^0.9.5"
django-allauth = "^0.52.0"

[tool.isort]
profile = "black"

[tool.flake8]
profile = "black"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pre-commit = "^3.0.4"
mypy = "^1.0.0"
pytest = "^7.2.1"
isort = "^5.12.0"
commitizen = "^2.41.0"

[tool.commitizen]
name = "cz_customize"
version = "0.1.1"
tag_format = "v$version"
update_changelog_on_bump = true
version_files = [
    "LinkCut/__init__.py:version",
    "pyproject.toml:^version"
]

[tool.commitizen.customize]
message_template = "{{change_type}}:{% if show_message %} {{message}}{% endif %}"
example = "feature: this feature enable customize through config file"
schema = "<type>: <body>"
bump_pattern = "^(major|minor|patch|hotfix)"
bump_map = {"major" = "MAJOR", "minor" = "MINOR", "patch" = "PATCH", "hotfix" = "PATCH"}

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = "C0114, C0115, R0903, C0116, C0415, W0614, E1101"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
